Remove popup fragment_income and fragment_popup, change to activity_popup.xml. fragment_income merge with activity_add_new_income.xml
