+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Jom_Finance">
         <activity
+            android:name=".income.DetailIncome"
+            android:exported="false" />
+        <activity
             android:name=".PopupActivity"
             android:exported="false" />
         <activity
