+++ app/src/main/java/com/example/jom_finance/income/AddNewIncome.kt
@@ -1,20 +1,34 @@
 package com.example.jom_finance.income
 
 import android.app.AlertDialog
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.LayoutInflater
-import androidx.fragment.app.FragmentTransaction
-import com.example.jom_finance.PopupActivity
+import android.widget.ArrayAdapter
 import com.example.jom_finance.R
-import com.example.jom_finance.SignUpActivity
-import kotlinx.android.synthetic.main.fragment_income.*
+import kotlinx.android.synthetic.main.activity_add_new_income.*
 
 class AddNewIncome : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_add_new_income)
 
+        ArrayAdapter.createFromResource(
+            this,
+            R.array.Category,
+            R.layout.spinner_list
+        ).also { adapter ->
+            // Specify the layout to use when the list of choices appears
+            adapter.setDropDownViewResource(R.layout.spinner_list)
+            // Apply the adapter to the spinner
+            spinnerCategory.adapter = adapter
+        }
+        AddnewBtn.setOnClickListener{
+            val resetView = LayoutInflater.from(this).inflate(R.layout.activity_popup, null)
+            val resetViewBuilder = AlertDialog.Builder(this,R.style.CustomAlertDialog).setView(resetView)
+            //show dialog
+            val displayDialog = resetViewBuilder.show()
+        }
     }
+
 }
