--- app/src/main/java/com/example/jom_finance/income/incomeFragment.kt
@@ -1,82 +1,0 @@
-package com.example.jom_finance.income
-
-import android.app.AlertDialog
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.jom_finance.R
-import android.widget.ArrayAdapter
-import com.example.jom_finance.PopupActivity
-import kotlinx.android.synthetic.main.fragment_income.*
-import kotlinx.android.synthetic.main.fragment_income.view.*
-
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [incomeFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class incomeFragment : Fragment() {
-        override fun onCreate(savedInstanceState: Bundle?) {
-         super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        val view = inflater.inflate(R.layout.fragment_income, container, false)
-        view.AddnewBtn.setOnClickListener{
-            val resetView = LayoutInflater.from(this.activity).inflate(R.layout.activity_popup, null)
-            val resetViewBuilder = AlertDialog.Builder(this.activity,R.style.CustomAlertDialog).setView(resetView)
-            //show dialog
-            val displayDialog = resetViewBuilder.show()
-        }
-        // Inflate the layout for this fragment
-        return view
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        ArrayAdapter.createFromResource(
-            activity as Context,
-            R.array.Category,
-            R.layout.spinner_list).also { adapter ->
-            // Specify the layout to use when the list of choices appears
-            adapter.setDropDownViewResource(R.layout.spinner_list)
-            // Apply the adapter to the spinner
-            spinnerCategory.adapter = adapter
-        }
-
-
-    }
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment incomeFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            incomeFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
-    }
-}
