+++ app/src/main/res/layout/activity_popup.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PopupActivity">
+    tools:context=".PopupActivity"
+    android:background="@color/white">
 
         <com.google.android.material.card.MaterialCardView
             android:id="@+id/popupCard"
@@ -20,7 +21,8 @@
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:padding="25dp">
+                android:padding="25dp"
+                android:background="@color/white">
 
                 <ImageView
                     android:id="@+id/imageView2"
