--- app/src/main/res/layout/fragment_income.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:fillViewport="true">
-
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".income.incomeFragment"
-    android:background="@drawable/rounded_corner"
-    android:backgroundTint="@color/white"
-    >
-<androidx.constraintlayout.widget.ConstraintLayout
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/CategoryCard"
-        android:layout_width="343dp"
-        android:layout_height="56dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="10dp"
-        app:cardCornerRadius="18dp"
-        android:backgroundTint="@color/white"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <androidx.appcompat.widget.AppCompatSpinner
-            android:id="@+id/spinnerCategory"
-            android:layout_width="match_parent"
-            android:layout_height="58dp"
-            android:spinnerMode="dropdown"
-            android:theme="@style/ThemeOverlay.AppCompat.Light" />
-    </androidx.cardview.widget.CardView>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:layout_width="343dp"
-        android:layout_height="56dp"
-        android:layout_marginTop="40dp"
-        android:id="@+id/DescriptionLayout"
-        app:boxStrokeWidth="0dp"
-        android:hint="Description"
-        android:textColorHint="#91919F"
-        app:boxStrokeWidthFocused="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/CategoryCard"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/DescriptionField"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:ems="15"
-            android:background="@drawable/custom_input"
-            android:singleLine="true"
-            android:textColor="@color/black"
-            android:padding="10dp"/>
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/wallet"
-        android:layout_width="343dp"
-        android:layout_height="56dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="10dp"
-        android:layout_marginTop="32dp"
-        app:cardCornerRadius="18dp"
-        android:backgroundTint="@color/white"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/DescriptionLayout">
-
-        <androidx.appcompat.widget.AppCompatSpinner
-            android:id="@+id/spinnerWallet"
-            android:layout_width="match_parent"
-            android:layout_height="58dp"
-            android:spinnerMode="dropdown"
-            android:theme="@style/ThemeOverlay.AppCompat.Light" />
-    </androidx.cardview.widget.CardView>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:layout_width="343dp"
-        android:layout_height="56dp"
-        android:layout_marginTop="40dp"
-        android:id="@+id/AttachmentLayout"
-        app:boxStrokeWidth="0dp"
-        android:hint="Add Attachment"
-        android:textColorHint="#91919F"
-        app:boxStrokeWidthFocused="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/wallet"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/attachment"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:ems="15"
-            android:background="@drawable/custom_input"
-            android:singleLine="true"
-            android:textColor="@color/black"
-            android:padding="10dp"/>
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.switchmaterial.SwitchMaterial
-        android:id="@+id/RepeatToggle"
-        android:layout_width="338dp"
-        android:layout_height="54dp"
-        android:layout_marginTop="32dp"
-        android:hint="Repeat"
-        android:textStyle="bold"
-        android:textColorHint="#292B2D"
-        android:textSize="16sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/AttachmentLayout">
-
-    </com.google.android.material.switchmaterial.SwitchMaterial>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/AddnewBtn"
-        android:layout_width="330dp"
-        android:layout_height="56dp"
-        android:layout_marginTop="40dp"
-        android:background="@drawable/rounded_corner"
-        android:backgroundTint="#5D5FEF"
-        android:text="Add"
-        android:textColor="@color/white"
-        android:textSize="17dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.493"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/RepeatToggle" />
-</androidx.constraintlayout.widget.ConstraintLayout>
-</FrameLayout>
-</ScrollView>
