+++ app/src/main/res/values/style.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="CustomAlertDialog" parent="Theme.AppCompat.Light.Dialog.Alert">
+    <style name="CustomAlertDialog" parent="Theme.AppCompat.Dialog.Alert">
       <item name="android:windowBackground">@drawable/popup_background</item>
     </style>
 </resources>
